third_party/blink/web_tests/fast/dom/shadow/shadow-root-js-api-expected.txt

Tests for ShadowRoot JS APIs. Can only run within DRT

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.ShadowRoot is defined.
PASS new window.ShadowRoot(document.createElement('div')) threw exception TypeError: Failed to construct 'ShadowRoot': Illegal constructor.
PASS shadowRoot instanceof window.ShadowRoot is true
PASS shadowRoot.getElementById('div1') is div1
PASS shadowRoot.getElementById('foo') is null
PASS shadowRoot.querySelector('#div1') is div1
PASS shadowRoot.querySelector('#foo') is null
PASS shadowRoot.querySelector('.class2') is div2
PASS shadowRoot.querySelector('.foo') is null
PASS shadowRoot.querySelectorAll('div') is [div1, div2]
PASS shadowRoot.querySelectorAll('foo') is []
PASS successfullyParsed is true

TEST COMPLETE
